FULL CODE REVIEW
Cleaned up the Java executor sources by removing a stray file, fixing the Redis client filename, and rewriting the truncated spread opportunity class. The updated class now provides placeholder parsing along with the necessary getters

Modified the near-miss logger to use doubles instead of BigDecimal when persisting spread opportunities

Documented the latest batch of fixes in the changelog, highlighting updates to executor files and logging behavior

+++ executor/src/main/java/NearMissLogger.java
@@ -21,10 +21,10 @@ public class NearMissLogger {
             stmt.setString(1, opp.getBuyExchange());
             stmt.setString(2, opp.getSellExchange());
             stmt.setString(3, opp.getPair());
-            stmt.setBigDecimal(4, opp.getGrossEdge());
-            stmt.setBigDecimal(5, opp.getNetEdge());
+            stmt.setDouble(4, opp.getGrossEdge());
+            stmt.setDouble(5, opp.getNetEdge());
             stmt.setString(6, reason);
-            stmt.setBigDecimal(7, opp.getExpectedProfitUsd());
+            stmt.setDouble(7, opp.getExpectedProfitUsd());
             stmt.setLong(8, opp.getLatencyMs());
             stmt.executeUpdate();
         } catch (SQLException e) {
